+++ tests/testthat/test-s3.R
@@ -83,3 +83,26 @@ test_that("can pass additional attributes to boxes", {
   box <- new_box(NA, "foo", bar = "baz")
   expect_identical(box %@% bar, "baz")
 })
+
+test_that("done() boxes values", {
+  expect_true(is_done_box(done(3)))
+  expect_identical(unbox(done(3)), 3)
+  expect_identical(done(3) %@% empty, FALSE)
+})
+
+test_that("done() can be empty", {
+  empty <- done()
+
+  expect_identical(unbox(empty), missing_arg())
+
+  expect_true(is_done_box(empty))
+  expect_is(empty, "rlang_box_done")
+  expect_identical(empty %@% empty, TRUE)
+
+  expect_true(is_done_box(empty, empty = TRUE))
+  expect_false(is_done_box(empty, empty = FALSE))
+
+  nonempty <- done(missing_arg())
+  expect_false(is_done_box(nonempty, empty = TRUE))
+  expect_true(is_done_box(nonempty, empty = FALSE))
+})