+++ NEWS.md
@@ -1,6 +1,11 @@
 
 # rlang 0.3.0.9000
 
+* New `done()` function to box a value. Done boxes are sentinels to
+  indicate early termination of a loop or computation. For instance,
+  it will be used in the purrr package to allow users to shortcircuit
+  a reduction or accumulation.
+
 * `new_box()` now accepts additional attributes passed to `structure()`.
 
 * `as_string()` now unwraps quosured symbols automatically.